Rename new::TEdgeConfig as TEdgeConfig
