+++ crates/core/tedge/src/cli/config/commands/set.rs
@@ -1,6 +1,6 @@
 use crate::command::Command;
-use tedge_config::new::WritableKey;
 use tedge_config::TEdgeConfigRepository;
+use tedge_config::WritableKey;
 
 pub struct SetConfigCommand {
     pub key: WritableKey,
