+++ crates/core/tedge/src/cli/connect/command.rs
@@ -12,8 +12,8 @@ use rumqttc::QoS::AtLeastOnce;
 use std::path::Path;
 use std::sync::Arc;
 use std::time::Duration;
-use tedge_config::new::TEdgeConfig;
 use tedge_config::system_services::*;
+use tedge_config::TEdgeConfig;
 use tedge_config::*;
 use tedge_utils::paths::create_directories;
 use tedge_utils::paths::ok_if_not_found;
