+++ plugins/c8y_configuration_plugin/src/main.rs
@@ -6,10 +6,10 @@ use clap::Parser;
 use std::path::Path;
 use std::path::PathBuf;
 use tedge_actors::Runtime;
-use tedge_config::new::TEdgeConfig;
 use tedge_config::system_services::get_log_level;
 use tedge_config::system_services::set_log_level;
 use tedge_config::CertificateError;
+use tedge_config::TEdgeConfig;
 use tedge_config::DEFAULT_TEDGE_CONFIG_PATH;
 use tedge_file_system_ext::FsWatchActorBuilder;
 use tedge_health_ext::HealthMonitorBuilder;
