+++ plugins/c8y_firmware_plugin/src/main.rs
@@ -4,9 +4,9 @@ use c8y_http_proxy::credentials::C8YJwtRetriever;
 use clap::Parser;
 use std::path::PathBuf;
 use tedge_actors::Runtime;
-use tedge_config::new::TEdgeConfig;
 use tedge_config::system_services::get_log_level;
 use tedge_config::system_services::set_log_level;
+use tedge_config::TEdgeConfig;
 use tedge_config::DEFAULT_TEDGE_CONFIG_PATH;
 use tedge_downloader_ext::DownloaderActor;
 use tedge_health_ext::HealthMonitorBuilder;
